+++ lib/committee/schema_validator.rb
@@ -4,7 +4,7 @@ module Committee
   module SchemaValidator
     class << self
       def request_media_type(request)
-        request.content_type.to_s.split(";").first.to_s
+        request.media_type.to_s
       end
 
       # @param [String] prefix