+++ test/committee_test.rb
@@ -31,18 +31,44 @@ describe Committee do
     end
   end
 
+  it "warns need_good_option" do
+    old_stderr = $stderr
+    $stderr = StringIO.new
+    begin
+      Committee.need_good_option "show"
+      assert_equal "show\n", $stderr.string
+    ensure
+      $stderr = old_stderr
+    end
+  end
+
   it "warns on deprecated unless $VERBOSE is nil" do
     old_stderr = $stderr
     old_verbose = $VERBOSE
     $stderr = StringIO.new
     begin
       $VERBOSE = nil
-      Committee.warn_deprecated "blah"
+      Committee.warn_deprecated_until_6 true, "blah"
       assert_equal "", $stderr.string
 
       $VERBOSE = true
-      Committee.warn_deprecated "blah"
+      Committee.warn_deprecated_until_6 true, "blah"
       assert_equal "[DEPRECATION] blah\n", $stderr.string
+    ensure
+      $stderr = old_stderr
+      $VERBOSE = old_verbose
+    end
+  end
+
+
+  it "doesn't warns on deprecated if cond is false" do
+    old_stderr = $stderr
+    old_verbose = $VERBOSE
+    $stderr = StringIO.new
+    begin
+      $VERBOSE = true
+      Committee.warn_deprecated_until_6 false, "blah"
+      assert_equal "", $stderr.string
     ensure
       $stderr = old_stderr
       $VERBOSE = old_verbose