+++ test/schema_validator/open_api_3/operation_wrapper_test.rb
@@ -32,12 +32,15 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
     ]
 
     it 'correct data' do
-      operation_object.validate_request_params(SCHEMA_PROPERTIES_PAIR.to_h, HEADER, @validator_option)
+      operation_object.validate_request_params({}, {}, SCHEMA_PROPERTIES_PAIR.to_h, HEADER, @validator_option)
       assert true
     end
 
     it 'correct object data' do
-      operation_object.validate_request_params({
+      operation_object.validate_request_params(
+                                  {},
+                                  {},
+                                  {
                                      "object_1" =>
                                          {
                                              "string_1" => nil,
@@ -54,7 +57,7 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
 
     it 'invalid params' do
       e = assert_raises(Committee::InvalidRequest) {
-        operation_object.validate_request_params({"string" => 1}, HEADER, @validator_option)
+        operation_object.validate_request_params({}, {}, {"string" => 1}, HEADER, @validator_option)
       }
 
       assert_match(/expected string, but received Integer: 1/i, e.message)
@@ -63,10 +66,10 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
 
     it 'support put method' do
       @method = "put"
-      operation_object.validate_request_params({"string" => "str"}, HEADER, @validator_option)
+      operation_object.validate_request_params({}, {}, {"string" => "str"}, HEADER, @validator_option)
 
       e = assert_raises(Committee::InvalidRequest) {
-        operation_object.validate_request_params({"string" => 1}, HEADER, @validator_option)
+        operation_object.validate_request_params({}, {}, {"string" => 1}, HEADER, @validator_option)
       }
 
       assert_match(/expected string, but received Integer: 1/i, e.message)
@@ -75,10 +78,10 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
 
     it 'support patch method' do
       @method = "patch"
-      operation_object.validate_request_params({"integer" => 1}, HEADER, @validator_option)
+      operation_object.validate_request_params({}, {}, {"integer" => 1}, HEADER, @validator_option)
 
       e = assert_raises(Committee::InvalidRequest) {
-        operation_object.validate_request_params({"integer" => "str"}, HEADER, @validator_option)
+        operation_object.validate_request_params({}, {}, {"integer" => "str"}, HEADER, @validator_option)
       }
 
       assert_match(/expected integer, but received String: "str"/i, e.message)
@@ -86,7 +89,7 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
     end
 
     it 'unknown param' do
-      operation_object.validate_request_params({"unknown" => 1}, HEADER, @validator_option)
+      operation_object.validate_request_params({}, {}, {"unknown" => 1}, HEADER, @validator_option)
     end
 
     describe 'support get method' do
@@ -96,13 +99,17 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
 
       it 'correct' do
         operation_object.validate_request_params(
+            {},
             {"query_string" => "query", "query_integer_list" => [1, 2]},
+            {},
             HEADER,
             @validator_option
         )
 
         operation_object.validate_request_params(
+            {},
             {"query_string" => "query", "query_integer_list" => [1, 2], "optional_integer" => 1},
+            {},
             HEADER,
             @validator_option
         )
@@ -112,7 +119,7 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
 
       it 'not exist required' do
         e = assert_raises(Committee::InvalidRequest) {
-          operation_object.validate_request_params({"query_integer_list" => [1, 2]}, HEADER, @validator_option)
+          operation_object.validate_request_params({}, {"query_integer_list" => [1, 2]}, {}, HEADER, @validator_option)
         }
 
         assert_match(/missing required parameters: query_string/i, e.message)
@@ -122,7 +129,9 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
       it 'invalid type' do
         e = assert_raises(Committee::InvalidRequest) {
           operation_object.validate_request_params(
+              {},
               {"query_string" => 1, "query_integer_list" => [1, 2], "optional_integer" => 1},
+              {},
               HEADER,
               @validator_option
           )
@@ -140,14 +149,14 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
       end
 
       it 'correct' do
-        operation_object.validate_request_params({"limit" => "1"}, HEADER, @validator_option)
+        operation_object.validate_request_params({}, {"limit" => "1"}, {}, HEADER, @validator_option)
 
         assert true
       end
 
       it 'invalid type' do
         e = assert_raises(Committee::InvalidRequest) {
-          operation_object.validate_request_params({"limit" => "a"}, HEADER, @validator_option)
+          operation_object.validate_request_params({}, {"limit" => "a"}, {}, HEADER, @validator_option)
         }
 
         assert_match(/expected integer, but received String: "a"/i, e.message)
@@ -162,14 +171,14 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
       end
 
       it 'correct' do
-        operation_object.validate_request_params({"limit" => "1"}, HEADER, @validator_option)
+        operation_object.validate_request_params({}, {"limit" => "1"}, {}, HEADER, @validator_option)
 
         assert true
       end
 
       it 'invalid type' do
         e = assert_raises(Committee::InvalidRequest) {
-          operation_object.validate_request_params({"limit" => "a"}, HEADER, @validator_option)
+          operation_object.validate_request_params({}, {"limit" => "a"}, {}, HEADER, @validator_option)
         }
 
         assert_match(/expected integer, but received String: "a"/i, e.message)
@@ -179,10 +188,10 @@ describe Committee::SchemaValidator::OpenAPI3::OperationWrapper do
 
     it 'support options method' do
       @method = "options"
-      operation_object.validate_request_params({"integer" => 1}, HEADER, @validator_option)
+      operation_object.validate_request_params({}, {}, {"integer" => 1}, HEADER, @validator_option)
 
       e = assert_raises(Committee::InvalidRequest) {
-        operation_object.validate_request_params({"integer" => "str"}, HEADER, @validator_option)
+        operation_object.validate_request_params({}, {}, {"integer" => "str"}, HEADER, @validator_option)
       }
 
       assert_match(/expected integer, but received String: "str"/i, e.message)